eth: move PrivateTxPoolAPI to api.go
This is a small addition and does not warrant
its own file. It fits well with the exisiting
API types and methods in the existing file.
